Refactoring of callback
