+++ core/src/main/scala/stainless/frontend/CallBackWithRegistry.scala
@@ -1,0 +1,122 @@
+/* Copyright 2009-2016 EPFL, Lausanne */
+
+package stainless
+package frontend
+
+import extraction.xlang.{ trees => xt }
+import utils.{ DependenciesFinder, Registry }
+
+import scala.collection.mutable.{ ListBuffer, Map => MutableMap }
+
+trait CallBackWithRegistry extends CallBack { self =>
+
+  /******************* Public Interface: Override CallBack ***************************************/
+
+  final override def beginExtractions(): Unit = { /* nothing */ }
+
+  final override def apply(file: String, unit: xt.UnitDef,
+                           classes: Seq[xt.ClassDef], functions: Seq[xt.FunDef]): Unit = {
+    ctx.reporter.debug(s"Got a unit for $file: ${unit.id} with:")
+    ctx.reporter.debug(s"\tfunctions -> [${functions.map { _.id }.sorted mkString ", "}]")
+    ctx.reporter.debug(s"\tclasses   -> [${classes.map { _.id }.sorted mkString ", "}]")
+
+    // Remove any node from the registry that no longer exists.
+    previousFileData get file foreach { case (prevClasses, prevFuns) =>
+      val removedClasses = prevClasses filterNot { cd => classes exists { _.id == cd.id } }
+      val removedFuns = prevFuns filterNot { cd => functions exists { _.id == cd.id } }
+      ctx.reporter.debug(s"Removing the following from the registry:")
+      ctx.reporter.debug(s"\tfunctions -> [${removedFuns.map { _.id }.sorted mkString ", "}]")
+      ctx.reporter.debug(s"\tclasses   -> [${removedClasses.map { _.id }.sorted mkString ", "}]")
+      registry.remove(removedClasses, removedFuns)
+    }
+
+    // Update our state with the new data, producing new symbols through the registry.
+    previousFileData += file -> (classes, functions)
+    val symss = registry.update(classes, functions)
+    processSymbols(symss)
+  }
+
+  final override def endExtractions(): Unit = {
+    val symss = registry.checkpoint()
+    processSymbols(symss)
+  }
+
+  final override def stop(): Unit = tasks foreach { _.cancel(true) }
+
+  final override def join(): Unit = tasks foreach { _.get }
+
+  // See assumption/requirements in [[CallBack]]
+  final override def getReports: Seq[Report] = tasks map { _.get } filter { _ != null }
+
+
+  /******************* Customisation Points *******************************************************/
+
+  protected type Report <: AbstractReport
+
+  protected val ctx: inox.Context
+
+  /** Produce a report for the given program, in a blocking fashion. */
+  protected def solve(program: Program { val trees: extraction.xlang.trees.type }): Report
+
+  /** Checks whether the given function/class should be processed at some point. */
+  protected def shouldBeChecked(fd: xt.FunDef): Boolean
+  protected def shouldBeChecked(cd: xt.ClassDef): Boolean
+
+
+  /******************* Internal State *************************************************************/
+
+  private implicit val debugSection = DebugSectionFrontend
+  private val tasks = ListBuffer[java.util.concurrent.Future[Report]]()
+
+  private val previousFileData = MutableMap[String, (Seq[xt.ClassDef], Seq[xt.FunDef])]()
+
+  private val registry = new Registry {
+    override val ctx = self.ctx
+
+    override def computeDirectDependencies(fd: xt.FunDef): Set[Identifier] = new DependenciesFinder()(fd)
+    override def computeDirectDependencies(cd: xt.ClassDef): Set[Identifier] = new DependenciesFinder()(cd)
+
+    override def shouldBeChecked(fd: xt.FunDef): Boolean = self.shouldBeChecked(fd)
+    override def shouldBeChecked(cd: xt.ClassDef): Boolean = self.shouldBeChecked(cd)
+  }
+
+
+  /******************* Internal Helpers ***********************************************************/
+
+  private def processSymbols(symss: Iterable[xt.Symbols]): Unit = symss foreach { syms =>
+    // The registry tells us something should be verified in these symbols.
+    val program = new inox.Program {
+      val trees: extraction.xlang.trees.type = extraction.xlang.trees
+      val ctx = self.ctx
+      val symbols = syms
+    }
+
+    try {
+      syms.ensureWellFormed
+    } catch {
+      case e: syms.TypeErrorException =>
+        ctx.reporter.error(e.pos, e.getMessage)
+        ctx.reporter.error(s"The extracted sub-program in not well formed.")
+        ctx.reporter.error(s"Symbols are:")
+        ctx.reporter.error(s"functions -> [${syms.functions.keySet.toSeq.sorted mkString ", "}]")
+        ctx.reporter.error(s"classes   -> [\n  ${syms.classes.values mkString "\n  "}\n]")
+        ctx.reporter.fatalError(s"Aborting from CallBackWithRegistry")
+    }
+
+    ctx.reporter.debug(s"Solving program with ${syms.functions.size} functions & ${syms.classes.size} classes")
+
+    processProgram(program)
+  }
+
+  private def processProgram(program: Program { val trees: extraction.xlang.trees.type }): Unit = {
+    // Dispatch a task to the executor service instead of blocking this thread.
+    val task = new java.util.concurrent.Callable[Report] {
+      override def call(): Report = solve(program)
+    }
+
+    val future = MainHelpers.executor.submit(task)
+    this.synchronized { tasks += future }
+    // task.call() // For debug, comment the two previous lines and uncomment this one.
+  }
+
+}
